backend/requirements.txt
flask
flask_sqlalchemy
python-dotenv
python-pptx
openai
Werkzeug
